Refactor: extract Provider from Firebase component.
