+++ src/components/Firebase/index.js
@@ -1,6 +1,5 @@
 import FirebaseContext from "./context";
 import Firebase from "./firebase";
-import Provider from "./provider";
-import { useAuthUser } from "./useAuthUser";
 
-export { Firebase, FirebaseContext, Provider, useAuthUser };
+export default Firebase;
+export { FirebaseContext };
