+++ src/components/Provider/index.js
@@ -1,0 +1,5 @@
+import Provider from "./provider";
+import { useAuthUser } from "./useAuthUser";
+
+export default Provider;
+export { useAuthUser };
